% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glfMB.R
\name{Blasius}
\alias{Blasius}
\title{Blasius correlation for the Darcy friction factor}
\usage{
Blasius(Re)
}
\arguments{
\item{Re}{Reynold number}
}
\value{
Darcy friction factor
}
\description{
Calculate the Darcy friction factor with assuming a smooth pipe
}
