% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glfMB_facade.R
\name{call_MB}
\alias{call_MB}
\title{Calculate flow regime, holdup, and pressure drop with the model of Mukherjee and Brill (1985)}
\usage{
call_MB(vsG, vsL, ID, densityG, densityL,
        viscosityG, viscosityL, surfaceTension,
        angle, roughness, pressure)
}
\arguments{
\item{vsG}{Superficial velocity of gas - m/s}

\item{vsL}{Superficial velocity of liquid - m/s}

\item{ID}{Pipe internal diameter - m}

\item{densityG}{Density of gas - kg/m3}

\item{densityL}{Density of liquid - kg/m3}

\item{viscosityG}{Visosity of gas - Pa-s}

\item{viscosityL}{Visosity of liquid - Pa-s}

\item{surfaceTension}{Surface tension - N/m}

\item{angle}{Pipe angle (a positive value means upward) - radian}

\item{roughness}{Pipe roughness - m}

\item{pressure}{Pressure (optional) - Pa}
}
\value{
Data frame including following data:
\itemize{
\item \code{fr}: Flow regime (1: Stratified, 2: Annular, 3: Slug, and 4: Bubbly)
\item \code{hl}: Holdup
\item \code{dPdL}: Pressure drop per unit length - Pa/m
\item \code{dPdL_H}: Pressure drop per unit length due to hydrostatic - Pa/m
\item \code{dPdL_F}: Pressure drop per unit length due to friction - Pa/m
\item \code{dPdL_A}: Pressure drop per unit length due to acceleration - Pa/m
}
}
\description{
Call core functions of Mukherjee & Brill model: \code{l_dlns_MB()}, \code{l_flow_regime_MB()},
\code{l_holdup_MB()}, and \code{l_dPdL_MB()} to calculate flow regime, holdup, and pressure drop.
The last argument (\code{pressure}) is for consideration of accelaration, which you can ignore.
}
\note{
You can execute the calculation step by step by calling low-level functions if you need. Below is an example.\preformatted{dlns <- l_dlns_MB(vsG, vsL, ID, densityG, densityL,
                  viscosityG, viscosityL, surfaceTension, angle)
fr <- l_flow_regime_MB(dlns)
hl <- l_holdup_MB(dlns, fr)
dPdL <- l_dPdL_MB(dlns, fr, hl, roughness, pressure, debug=FALSE)
}
}
\examples{
# This example is from Brill and Mukherjee (1999)
# "Multiphase Flow in Wells" p.21, p.46 (Example 3.2 and 4.8)
vsG <- 3.86 * 0.3048           # 3.86 ft/s
vsL <- 3.97 * 0.3048           # 3.97 ft/s
ID <- 6 * 0.0254               # 6 inch
densityG <- 5.88 * 16.01845    # 5.88 lbm/ft3
densityL <- 47.61 * 16.01845   # 47.61 lbm/ft3
viscosityG <- 0.016 / 1000     # 0.016 cp
viscosityL <- 0.97  / 1000     # 0.970 cp
surfaceTension <- 8.41 / 1000  # 8.41 dynes/cm
angle <- pi/2                  # 90 deg
roughness <- 0.00006 * 0.3048  # 0.00006 ft
pressure <- 1700 * 6894.76     # 1700 psia

# Results should be 3 (slug), 0.560, and about 4727 Pa/m (= 0.209 psi/ft)
call_MB(vsG, vsL, ID, densityG, densityL,
        viscosityG, viscosityL, surfaceTension,
        angle, roughness, pressure)

}
\references{
\itemize{
\item Mukherjee, H., and J. P. Brill. 1985. Pressure Drop Correlations for Inclined Two-Phase Flow. Journal of Energy Resources Technology, Transactions of the ASME 107 (4)
\item Mukherjee, H., and J. P. Brill. 1985. Empirical Equations to Predict Flow Patterns in Two-Phase Inclined Flow. International Journal of Multiphase Flow 11 (3)
\item J. P. Brill., and Mukherjee, H. 1999. Multiphase Flow in Wells. Society of Petroleum Engineers.
}
}
