% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glfMB.R
\name{Colebrook}
\alias{Colebrook}
\title{Colebrook correlation for the Darcy friction factor}
\usage{
Colebrook(roughness, D, Re, tol = 1e-08, itMax = 10, warn = TRUE)
}
\arguments{
\item{roughness}{Pipe roughness}

\item{D}{Pipe diameter}

\item{Re}{Reynold number}

\item{tol}{Tolerance in Newton-Raphson method (optional)}

\item{itMax}{Maximum number of iteration  (optional)}

\item{warn}{show warnings when Re <= 4000  (optional)}
}
\value{
Darcy friction factor
}
\description{
Calculate the Darcy friction factor with assuming a smooth pipe
}
