% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glfMB.R
\name{dlns_MB}
\alias{dlns_MB}
\title{Low-level function to calculate dimensionless numbers}
\usage{
dlns_MB(vsG, vsL, D, densityG, densityL,
        viscosityG, viscosityL, surfaceTension, angle)
}
\arguments{
\item{vsG}{Superficial velocity of gas}

\item{vsL}{Superficial velocity of liquid}

\item{D}{Pipe diameter}

\item{densityG}{Density of gas}

\item{densityL}{Density of liquid}

\item{viscosityG}{Visosity of gas}

\item{viscosityL}{Visosity of liquid}

\item{surfaceTension}{Surface tension}

\item{angle}{Pipe angle in radian (0 is horizontal flow)}
}
\value{
Data frame of dimensionless numbers (\code{NLv}, \code{NGv}, \code{Nd}, \code{NL}, \code{NGvSM}, \code{NLvBS}, \code{NGvBS_up}, and \code{NLvST}) and input values
}
\description{
Return the dimensionless groups proposed by Duns & Ros:
\itemize{
\item \code{NLv}: Liquid velocity number
\item \code{NGv}: Gas velocity number
\item \code{Nd}: Pipe diameter number
\item \code{NL}: Liquid viscosity number
\item \code{NGvSM}: Slug/(Annular Mist) transition
\item \code{NGvBS}: Bubble/Slug transition (Downflow)
\item \code{NLvBS_up}: Bubble/Slug transition (Upflow)
\item \code{NLvST}: Stratified (Downflow)
}
}
\examples{
\dontrun{
vsG <- 3.86 * 0.3048    # 3.86 ft/s
vsL <- 3.97 * 0.3048    # 3.97 ft/s
D   <- 6 * 0.0254       # 6 inch
densityG <- 5.88 * 16.01845     # 5.88 lbm/ft3  - (1 lbm/ft3 = 16.01845 kg/m3)
densityL <- 47.61 * 16.01845    # 47.61 lbm/ft3 - (1 lbm/ft3 = 16.01845 kg/m3)
viscosityG <- 0.016 / 1000      # 0.016 cp
viscosityL <- 0.97  / 1000      # 0.970 cp
surfaceTension <- 8.41 / 1000   # 8.41 dynes/cm
angle <- pi/2                   # 90 deg

#  NLv=11.84, NGv=11.54, NGvSM=350.8, NLvBS_up=18.40, NL=0.0118
dlns_MB(vsG, vsL, D, densityG, densityL, viscosityG, viscosityL, surfaceTension, angle)
}
}
