% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glfMB.R
\name{dPdL_MB}
\alias{dPdL_MB}
\title{Low-level function to calculate pressure drop (dP/L)}
\usage{
dPdL_MB(DLNs, flowRegime, HL, roughness, pressure, debug = FALSE)
}
\arguments{
\item{DLNs}{dimensionless numbers calculated by \code{dlns_MB()}}

\item{flowRegime}{flow regime estimated by \code{flow_regime_MB()}}

\item{HL}{Holdup}

\item{roughness}{Pipe roughness}

\item{pressure}{Pressure (optional)}

\item{debug}{If TRUE, print parameters (usually set to FALSE)}
}
\value{
pressure drop (Pa/m)
}
\description{
Low-level function to calculate pressure drop (dP/L)
}
