% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glfMB.R
\name{flow_regime_MB}
\alias{flow_regime_MB}
\title{Determine flow regime}
\usage{
flow_regime_MB(DLNs)
}
\arguments{
\item{DLNs}{Dimensionless numbers calculated by \code{dlns_MB()}}
}
\value{
a vector of numbers indicating flow regime (1: Stratified, 2: Annular, 3: Slug, and 4: Bubbly)
}
\description{
There are four types of defined flow regime:
\itemize{
\item 1: Stratified
\item 2: Annular
\item 3: Slug
\item 4: Bubbly
}
}
