% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glfMB.R
\name{generate_frm_MB}
\alias{generate_frm_MB}
\title{Generate flow regime map (frm) data}
\usage{
generate_frm_MB(
  vector_vsG,
  vector_vsL,
  D,
  densityG,
  densityL,
  viscosityG,
  viscosityL,
  surfaceTension,
  angle
)
}
\arguments{
\item{vector_vsG}{Vector of Superficial velocity of gas - m/s}

\item{vector_vsL}{Vector of superficial velocity of liquid - m/s}

\item{D}{Pipe diameter - m}

\item{densityG}{Density of gas - kg/m3}

\item{densityL}{Density of liquid - kg/m3}

\item{viscosityG}{Visosity of gas - Pa-s}

\item{viscosityL}{Visosity of liquid - Pa-s}

\item{surfaceTension}{Surface tension - N/m}

\item{angle}{Pipe angle (0 is horizontal flow) - radian}
}
\value{
A matrix (class="frm_MB") including the information of flow regime at
specified superfical velocities of gas and liquid.
\itemize{
\item \code{vsG}: Superficial velocity of gas
\item \code{vsL}: Superficial velocity of liquid
\item \code{fr}: Flow regime (1: Stratified, 2: Annular, 3: Slug, and 4: Bubbly)
\item \code{NGv}, \code{NLv}, \code{NL}, \code{NGvSM}, \code{NGvBS}, \code{NLvBS_up}, \code{NLvST}: Properties used in model
}
}
\description{
Generate data (matrix) for flow regime map. Range of the map is specified by \code{vector_vsG} and \code{vector_vsL}.
}
