% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glfMB.R
\name{holdup_MB}
\alias{holdup_MB}
\title{Calculate holdup}
\usage{
holdup_MB(DLNs, flowRegime)
}
\arguments{
\item{DLNs}{dimensionless numbers calculated by \code{dlns_MB()}}

\item{flowRegime}{flow regime estimated by \code{flow_regime_MB()}}
}
\value{
holdup
}
\description{
Calculate holdup
}
