% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MukherjeeBrill.R
\name{l_Darcy_friction_factor}
\alias{l_Darcy_friction_factor}
\title{Low-level function to calculate the Darcy friction factor}
\usage{
l_Darcy_friction_factor(Re, roughness, D, tol = 1e-08, itMax = 10)
}
\arguments{
\item{Re}{Reynolds number}

\item{roughness}{Pipe roughness}

\item{D}{Pipe diameter}

\item{tol}{Tolerance in \code{util_MB_Colebrook()} (optional)}

\item{itMax}{Maximum number of iteration in \code{util_MB_Colebrook()} (optional)}
}
\value{
Darcy friction factor
}
\description{
Calculate the Darcy friction factor considering Reynolds number and pipe roughness.
\itemize{
\item Re <= 2000: Laminar flow (= 64/Re)
\item Re >= 4000: Turbulent flow (Colebrook correlation)
\item 2000 < Re < 4000: Transition (see details)
}
}
\seealso{
\code{\link[=util_MB_Blasius]{util_MB_Blasius()}}, \code{\link[=util_MB_Colebrook]{util_MB_Colebrook()}}
}
