% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glfMB.R
\name{l_dPdL_MB}
\alias{l_dPdL_MB}
\title{Low-level function to calculate pressure drop}
\usage{
l_dPdL_MB(DLNs, flowRegime, HL, roughness, pressure, debug = FALSE)
}
\arguments{
\item{DLNs}{dimensionless numbers calculated by \code{l_dlns_MB()}}

\item{flowRegime}{flow regime estimated by \code{l_flow_regime_MB()}}

\item{HL}{Holdup}

\item{roughness}{Pipe roughness}

\item{pressure}{Pressure (optional)}

\item{debug}{If TRUE, print parameters (usually set to FALSE)}
}
\value{
A vector (or matrix) including following properties (or columns):
\itemize{
\item ['dPdL']: pressure drop (Pa/m)
\item ['dPdL_H']: pressure drop due to hydrostatic (Pa/m)
\item ['dPdL_F']: pressure drop due to friction (Pa/m)
\item ['dPdL_A']: pressure drop due to acceleration  (Pa/m)
}
}
\description{
Low-level function to calculate pressure drop
}
