% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MukherjeeBrill.R
\name{l_holdup_MB}
\alias{l_holdup_MB}
\title{Low-level function to calculate holdup}
\usage{
l_holdup_MB(DLNs, flowRegime)
}
\arguments{
\item{DLNs}{dimensionless numbers calculated by \code{l_dlns_MB()}}

\item{flowRegime}{flow regime estimated by \code{l_flow_regime_MB()}}
}
\value{
holdup
}
\description{
Low-level function to calculate holdup
}
