% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glfMB.R
\name{plot.frm_MB}
\alias{plot.frm_MB}
\title{Plot a flow regime map}
\usage{
\method{plot}{frm_MB}(x, xlab, ylab, xval = "vsG", yval = "vsL", ...)
}
\arguments{
\item{x}{Flow regime map data created by `generate_frm_MB()`}

\item{xlab}{label x (optional)}

\item{ylab}{label y (optional)}

\item{xval}{'vsG' or 'NGv' (optional)}

\item{yval}{'vsL' or 'NLv' (optional)}

\item{...}{graphical parameters to plot}
}
\description{
Plot a flow regime map
}
\examples{
# plot(a)

}
