% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glfMB_facade.R
\name{testdata_MB}
\alias{testdata_MB}
\title{Get a list of functions for test (and example)}
\usage{
testdata_MB()
}
\value{
a list of functions for fluid properties. SI unit is used in the functions.
\itemize{
\item Kerosene_surfaceTension(temperature) - K
\item Kerosene_density(temperature) - K
\item Kerosene_viscosity(temperature) - K
\item Lubricating_surfaceTension(temperature) - K
\item Lubricating_density(temperature) - K
\item Lubricating_viscosity(temperature) - K
\item Air_density(temperature, pressure) - K, Pa
\item Air_viscosity(temperature, pressure) - K, Pa
}
}
\description{
This function returns a list of functions for fluid properties (see also Value).
}
\examples{
td <- testdata_MB()
td$Kerosene_viscosity(283.15)     # Kerosene Viscosity at 10 degC
td$Air_density(293.15, 5*100000)  # Air density at 20 degC and 5 bar

}
