% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MukherjeeBrill_facade.R
\name{util_MB_Blasius}
\alias{util_MB_Blasius}
\title{Blasius correlation for the Darcy friction factor}
\usage{
util_MB_Blasius(Re)
}
\arguments{
\item{Re}{Reynold number}
}
\value{
Darcy friction factor
}
\description{
Calculate the Darcy friction factor with Blasius correlation, assuming a smooth pipe.
}
