% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MukherjeeBrill_facade.R
\name{util_MB_vs_vector}
\alias{util_MB_vs_vector}
\title{Generate a vector of superficial velocities}
\usage{
util_MB_vs_vector(from, to, num_points, log_scale)
}
\arguments{
\item{from}{minimum value of superficial velocities}

\item{to}{maximum value of superficial velocities}

\item{num_points}{number of data points of the returned vector}

\item{log_scale}{If TRUE, log scale is used for intervals of data points (optional)}
}
\value{
a vector of superficial velocities
}
\description{
Generate a vector of superficial velocities
}
\examples{
util_MB_vs_vector(1, 100, 5)
# 1  25.75  50.50  75.25  100
util_MB_vs_vector(1, 100, 5, TRUE)
# 1  3.162278  10  31.622777  100

}
