% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glfMB_facade.R
\name{vs_vector_MB}
\alias{vs_vector_MB}
\title{Generate a vector of superficial velocities (vs)}
\usage{
vs_vector_MB(from, to, num_points, log_scale)
}
\arguments{
\item{from}{minimum value of superficial velocities}

\item{to}{maximum value of superficial velocities}

\item{num_points}{number of data points of the returned vector}

\item{log_scale}{a}
}
\description{
Generate a vector of superficial velocities (vs)
}
