% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glfMB.R
\name{dlns_MB}
\alias{dlns_MB}
\title{Calculate dimensionless numbers}
\usage{
dlns_MB(vsG, vsL, D, densityG, densityL,
        viscosityG, viscosityL, surfaceTension, angle)
}
\arguments{
\item{vsG}{Superficial velocity of gas}

\item{vsL}{Superficial velocity of liquid}

\item{D}{Pipe diameter}

\item{densityG}{Density of gas}

\item{densityL}{Density of liquid}

\item{viscosityG}{Visosity of gas}

\item{viscosityL}{Visosity of liquid}

\item{surfaceTension}{Surface tension}

\item{angle}{Pipe angle in radian (0 is horizontal flow)}
}
\value{
Data frame of dimensionless numbers (\code{NLv}, \code{NGv}, \code{Nd}, \code{NL}, \code{NGvSM}, \code{NLvBS}, \code{NGvBS_up}, and \code{NLvST}) and input values
}
\description{
Return the dimensionless groups proposed by Duns & Ros:
\itemize{
\item \code{NLv}: Liquid velocity number
\item \code{NGv}: Gas velocity number
\item \code{Nd}: Pipe diameter number
\item \code{NL}: Liquid viscosity number
\item \code{NGvSM}: Slug/(Annular Mist) transition
\item \code{NGvBS}: Bubble/Slug transition (Downflow)
\item \code{NLvBS_up}: Bubble/Slug transition (Upflow)
\item \code{NLvST}: Stratified (Downflow)
}
}
\examples{
\dontrun{
dlns_MB(1.18, 1.21, 0.152, 94.2, 762.6, 1.6e-05, 0.00097, 0.00841, 1.5708)
}
}
