% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glfMB.R
\name{exec_glfMB}
\alias{exec_glfMB}
\title{Execute Mukherjee & Brill model}
\usage{
exec_glfMB(vsG, vsL, D, densityG, densityL,
           viscosityG, viscosityL, surfaceTension,
           angle, pressure, roughness)
}
\arguments{
\item{vsG}{Superficial velocity of gas - m/s}

\item{vsL}{Superficial velocity of liquid - m/s}

\item{D}{Pipe diameter - m}

\item{densityG}{Density of gas - kg/m3}

\item{densityL}{Density of liquid - kg/m3}

\item{viscosityG}{Visosity of gas - Pa-s}

\item{viscosityL}{Visosity of liquid - Pa-s}

\item{surfaceTension}{Surface tension - N/m}

\item{angle}{Pipe angle (0 is horizontal flow) - radian}

\item{pressure}{Pressure - Pa}

\item{roughness}{Pipe roughness}
}
\value{
Data frame including following data:
\itemize{
\item \code{fr}: Flow regime (1: Stratified, 2: Annular, 3: Slug, and 4: Bubbly)
\item \code{hl}: Holdup
\item \code{dPdL}: Pressure drop per unit length - Pa/m
\item \code{NLv}: Liquid velocity number
\item \code{NGv}: Gas velocity number
\item \code{Nd}: Pipe diameter number
\item \code{NL}: Liquid viscosity number
\item \code{NGvSM}: Slug/(Annular Mist) transition
\item \code{NGvBS}: Bubble/Slug transition (Downflow)
\item \code{NLvBS_up}: Bubble/Slug transition (Upflow)
\item \code{NLvST}: Stratified (Downflow)
}
}
\description{
Calculate flow regime, holdup, and pressure drop with the model of Mukherjee and Brill (1985).
}
\note{
You can execute the calculation step by step without this function if want. Below is an example.\preformatted{  dlns <- dlns_MB(vsG, vsL, D, densityG, densityL, viscosityG, viscosityL, surfaceTension, angle)
  fr   <- flow_regime_MB(dlns)
  hl <- holdup_MB(dlns, fr)
  dPdL <- dPdL_MB(dlns, fr, hl, pressure)
}
}
\examples{
\dontrun{
  exec_glfMB(1.18, 1.21, 0.152, 94.2, 762.6, 1.6e-05, 0.00097, 0.00841, 1.5708)
}

}
\references{
\itemize{
\item Mukherjee, H., and J. P. Brill. 1985. Pressure Drop Correlations for Inclined Two-Phase Flow. Journal of Energy Resources Technology, Transactions of the ASME 107 (4)
\item Mukherjee, Hemanta, and James P. Brill. 1985. Empirical Equations to Predict Flow Patterns in Two-Phase Inclined Flow. International Journal of Multiphase Flow 11 (3)
}
}
