% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MukherjeeBrill_facade.R
\name{generate_frm_MB}
\alias{generate_frm_MB}
\title{Generate data for flow regime map (frm_MB)}
\usage{
generate_frm_MB(
  vector_vsG,
  vector_vsL,
  D,
  densityG,
  densityL,
  viscosityG,
  viscosityL,
  surfaceTension,
  angle
)
}
\arguments{
\item{vector_vsG}{Vector of Superficial velocity of gas - m/s}

\item{vector_vsL}{Vector of superficial velocity of liquid - m/s}

\item{D}{Pipe diameter - m}

\item{densityG}{Density of gas - kg/m3}

\item{densityL}{Density of liquid - kg/m3}

\item{viscosityG}{Visosity of gas - Pa-s}

\item{viscosityL}{Visosity of liquid - Pa-s}

\item{surfaceTension}{Surface tension - N/m}

\item{angle}{Pipe angle (0 is horizontal flow) - radian}
}
\value{
A matrix (class="frm_MB") including the information of flow regime (columns) at
specified superfical velocities of gas and liquid.
\itemize{
\item \code{vsG}: Superficial velocity of gas
\item \code{vsL}: Superficial velocity of liquid
\item \code{fr}: Flow regime (1: Stratified, 2: Annular, 3: Slug, and 4: Bubbly)
\item \code{NGv}, \code{NLv}, \code{NL}, \code{NGvSM}, \code{NGvBS}, \code{NLvBS_up}, \code{NLvST}: Internal properties used in the prediction of flow regime
}
}
\description{
Generate a matrix (class="frm_MB") for a flow regime map. Range of the map is specified by \code{vector_vsG} and \code{vector_vsL}.
}
\examples{
vs_range = glfMB::vs_vector_MB(0.1, 10, 20, TRUE)
frm <- generate_frm_MB(vs_range, vs_range, 0.1,
                       40, 1002, 1.1E-05, 1.6E-03, 0.0695, pi/2)
plot(frm)    # glfMB:::plot.frm_MB(frm)

}
\references{
Mukherjee, H., and J. P. Brill. 1985. Empirical Equations to Predict Flow Patterns in Two-Phase Inclined Flow. International Journal of Multiphase Flow 11 (3)
}
