% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MukherjeeBrill.R
\name{l_Darcy_friction_factor}
\alias{l_Darcy_friction_factor}
\title{Low-level function to calculate the Darcy friction factor}
\usage{
l_Darcy_friction_factor(Re, roughness, D, tol = 1e-08, itMax = 10)
}
\arguments{
\item{Re}{Reynold number}

\item{roughness}{Pipe roughness}

\item{D}{Pipe diameter}

\item{tol}{Tolerance in `Colebrook()` (optional)}

\item{itMax}{Maximum number of iteration in `Colebrook()` (optional)}
}
\value{
Darcy friction factor
}
\description{
Calculate the Darcy friction factor considering pipe roughness.
}
