% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MukherjeeBrill.R
\docType{package}
\name{MukherjeeBrill}
\alias{MukherjeeBrill}
\title{Mukherjee & Brill model for gas-liquid two phase flow in a circular pipe}
\description{
This R package provides functions to calculate flow regime, liquid holdup,
and pressure drop with the empirical model of Mukherjee & Brill (1985).
}
\references{
\itemize{
\item Mukherjee, H., and Brill J. P. 1985. Pressure Drop Correlations for Inclined Two-Phase Flow. Journal of Energy Resources Technology, Transactions of the ASME 107 (4)
\item Mukherjee, H., and Brill J. P. 1985. Empirical Equations to Predict Flow Patterns in Two-Phase Inclined Flow. International Journal of Multiphase Flow 11 (3)
\item Mukherjee, H., and Brill J. P. 1983. Liquid Holdup Correlations for Inclined Two-Phase Flow. JPT, Journal of Petroleum Technology 35(5):1003–8.
}
}
\author{
Shunsuke S
}
