% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MukherjeeBrill_facade.R
\name{plot.frm_MB}
\alias{plot.frm_MB}
\title{Plot a flow regime map}
\usage{
\method{plot}{frm_MB}(x, xlab, ylab, xval = "vsG", yval = "vsL", ...)
}
\arguments{
\item{x}{Flow regime map data created by \code{generate_frm_MB()}}

\item{xlab}{label x (optional)}

\item{ylab}{label y (optional)}

\item{xval}{'vsG' or 'NGv' (optional)}

\item{yval}{'vsL' or 'NLv' (optional)}

\item{...}{graphical parameters to plot}
}
\description{
This function is automatically called by plot() when the class is \code{frm_MB}.
}
\examples{
vs_range = glfMB::vs_vector_MB(0.1, 10, 20, TRUE)
frm <- generate_frm_MB(vs_range, vs_range, 0.1,
                       40, 1002, 1.1E-05, 1.6E-03, 0.0695, pi/2)
plot(frm)    # glfMB:::plot.frm_MB(frm)

}
\references{
Mukherjee, H., and J. P. Brill. 1985. Empirical Equations to Predict Flow Patterns in Two-Phase Inclined Flow. International Journal of Multiphase Flow 11 (3)
}
