% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MukherjeeBrill_facade.R
\docType{data}
\name{testdata_MB}
\alias{testdata_MB}
\title{List of functions and constants for test (and example)}
\format{
An object of class \code{list} of length 10.
}
\usage{
testdata_MB
}
\description{
A list containing functions and constants for fluid properties used in tests (see also Value).
SI unit is used.
\itemize{
\item Mair: Molar mass of air - kg/mol
\item Sutherland_vis0_air: Viscosity of air at 273 K - Pa-s
\item Sutherland_T0_air - K
\item Sutherland_C_air - K
\item ideal_gas_density(temperature, pressure, M) - kg/m3 (K, Pa)
\item Sutherland_viscosityG(temperature, viscosity0, t0, const) - Pa-s (K, Pa-s, K, K)
\item Kerosene_surfaceTension(temperature): K
\item Kerosene_density(temperature) - kg/m3 (K, Pa)
\item Kerosene_viscosity(temperature): K
}
}
\examples{
testdata_MB$Kerosene_viscosity(283.15)     # Kerosene Viscosity at 10 degC

density_air <- function(temperature, pressure) {
  testdata_MB$ideal_gas_density(temperature, pressure, testdata_MB$Mair)
} 
density_air(293.15, 5*100000)  # Air density at 20 degC and 5 bar

}
\keyword{datasets}
