% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MukherjeeBrill_facade.R
\name{vs_vector_MB}
\alias{vs_vector_MB}
\title{Generate a vector of superficial velocities}
\usage{
vs_vector_MB(from, to, num_points, log_scale)
}
\arguments{
\item{from}{minimum value of superficial velocities}

\item{to}{maximum value of superficial velocities}

\item{num_points}{number of data points of the returned vector}

\item{log_scale}{If TRUE, log scale is used for intervals of data points (optional)}
}
\value{
a vector of superficial velocities
}
\description{
Generate a vector of superficial velocities
}
\examples{
vs_vector_MB(1, 100, 5)
# 1  25.75  50.50  75.25  100
vs_vector_MB(1, 100, 5, TRUE)
# 1  3.162278  10  31.622777  100

}
